+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-07-07T09:35:41.014Z
+Generated: 2026-07-08T08:22:43.387Z
 
 - 2026-07-05 • Microsoft Defender for Cloud • Deprecation • Discovery and posture for serverless container workloads is now generally available
 - 2026-07-01 • Microsoft Defender for Cloud • Deprecation • Plan enablement API now blocks onboarding to five deprecated Defender plans
@@ -14,6 +14,7 @@ Generated: 2026-07-07T09:35:41.014Z
 - 2026-07-01 • Microsoft Sentinel • Action required • Call to action: update queries and automation by July 1, 2026 - standardized account entity naming in incidents and alerts
 - 2026-07-01 • Microsoft Sentinel • GA • , Microsoft Sentinel will be supported in the Defender portal only, and any remaining customers using the Azure portal will be automatically redirected.
 - 2026-07-01 • Microsoft Sentinel • Preview • Custom detection rules support in Microsoft Sentinel Repositories (Preview)
+- 2026-07-01 • Security Exposure Management • Ikke oppgitt • OT data connectors in Microsoft Security Exposure Management
 - 2026-07-01 • Unified SecOps • Deprecation • , Microsoft Sentinel will be supported in the Defender portal only, and any remaining customers using the Azure portal will be automatically redirected.
 - 2026-06-30 • Microsoft Defender for Cloud • GA • Expanded multicloud security coverage is now generally available
 - 2026-06-30 • Microsoft Defender for Cloud • GA • General availability of Defender for Key Vault in Azure Government cloud
@@ -21,4 +22,3 @@ Generated: 2026-07-07T09:35:41.014Z
 - 2026-06-18 • Microsoft Defender for Cloud • GA • Cloud security reporting is now generally available
 - 2026-06-17 • Microsoft Defender for Cloud • GA • API security posture management for Function Apps and Logic Apps is now generally available
 - 2026-06-15 • Microsoft Sentinel • Action required • Call to action: update older Microsoft Sentinel content as code (Sentinel repositories) API versions before June 15, 2026
-- 2026-06-08 • Microsoft Defender for Cloud • Preview • New multicloud security recommendations now in public preview